feat: add transition animation for date changes
- Add fade-in effect for workout component for smooth transition
- Add delay of 800ms to progress to reduce unnecessary circle flicker
- Add separate workoutList component and move suspense location
  - Only workouts are asynchronously changed, thus need to be suspended
