+++ frontend/src/common/CenteredProgress.tsx
@@ -1,10 +1,19 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import { Box, CircularProgress } from '@material-ui/core';
 
 function CenteredProgress() {
+  const [delayFinished, setDelayFinished] = useState(false);
+
+  useEffect(() => {
+    const timer = setTimeout(() => {
+      setDelayFinished(true);
+    }, 800);
+    return () => clearTimeout(timer);
+  }, []);
+
   return (
-    <Box flex={1} display="flex" flexDirection="row" justifyContent="center" alignItems="center">
-      <CircularProgress size={100}/>
+    <Box pt={5} display="flex" flexDirection="row" justifyContent="center" alignItems="center">
+      {delayFinished && <CircularProgress size={100}/>}
     </Box>
   );
 }
